+++ src/aspire/volume/__init__.py
@@ -6,7 +6,7 @@ from aspire.utils import ensure
 from aspire.utils.coor_trans import grid_2d
 from aspire.utils.fft import centered_fft2, centered_ifft2
 from aspire.utils.matlab_compat import m_flatten, m_reshape
-from aspire.utils.preprocess_C import downsample
+from aspire.utils.preprocess import downsample
 
 
 class Volume:
